+++ src/modify/limegradient.betterachievements/BetterAchievementLayer.cpp
@@ -1,7 +1,0 @@
-#include <Geode/Geode.hpp>
-#include "../../SwelvyBG.hpp"
-#include <alphalaneous.alphas_geode_utils/include/ObjectModify.hpp>
-
-using namespace geode::prelude;
-ADD_TAG("limegradient.betterachievements-BetterAchievementLayer");
-SET_SWELVY(BetterAchievementLayer, "limegradient.betterachievements/BetterAchievementLayer", "content-background");