+++ src/modify/omgrod.garage_plus/GPKofiLayer.cpp
@@ -1,7 +1,0 @@
-#include <Geode/Geode.hpp>
-#include "../../SwelvyBG.hpp"
-#include <alphalaneous.alphas_geode_utils/include/ObjectModify.hpp>
-
-using namespace geode::prelude;
-ADD_TAG("omgrod.garage_plus-GPKofiLayer");
-SET_SWELVY(GPKofiLayer, "omgrod.garage_plus/GPKofiLayer", "background");