+++ src/modify/omgrod.geometry-dash-surge/GDSBasementLayer.cpp
@@ -1,7 +1,0 @@
-#include <Geode/Geode.hpp>
-#include "../../SwelvyBG.hpp"
-#include <alphalaneous.alphas_geode_utils/include/ObjectModify.hpp>
-
-using namespace geode::prelude;
-ADD_TAG("omgrod.geometry-dash-surge-GDSBasementLayer");
-SET_SWELVY_SPRITE(GDSBasementLayer, "omgrod.geometry-dash-surge/GDSBasementLayer");