+++ src/modify/thesillydoggo.newgrounds_explorer/NewgroundsBrowserLayer.cpp
@@ -1,7 +1,0 @@
-#include <Geode/Geode.hpp>
-#include "../../SwelvyBG.hpp"
-#include <alphalaneous.alphas_geode_utils/include/ObjectModify.hpp>
-
-using namespace geode::prelude;
-ADD_TAG("thesillydoggo.newgrounds_explorer-NewgroundsBrowserLayer");
-SET_SWELVY(NewgroundsBrowserLayer, "thesillydoggo.newgrounds_explorer/NewgroundsBrowserLayer", "background");